requirements.txt
langchain 
openai 
chromadb 
langchain-experimental
unstructured
pillow 
pydantic 
lxml 
pillow 
matplotlib 
chromadb 
tiktoken
poppler-utils
open-flamingo
torch
tesseract
transformers
accelerate
bitsandbytes
llava-torch==1.1.1
InstructorEmbedding
sentence_transformers
langchain_core